refactor(agentspan): extract WorkflowTaskRefs from AgentCompiler
Move ensureUniqueRefNames + its dedup/update/replace helpers — a
general-purpose WorkflowTask ref-renaming algorithm called from
compile()'s shared post-processing step for every dispatch path, not
just graph-structure compilation — into their own package-private
class. No test files touched.
